+++ pelutils/ticktock.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
+from copy import deepcopy
 from time import perf_counter
+from typing import Iterable
 
 from pelutils import thousand_seps
 from pelutils.format import Table
@@ -57,6 +59,19 @@ class Profile:
         return len(self._hits)
 
 
+class _ProfileContext:
+
+    def __init__(self, tt, profile_name: str):
+        self.tt = tt
+        self.profile_name = profile_name
+
+    def __enter__(self):
+        ...
+
+    def __exit__(self, *args):
+        self.tt.end_profile(self.profile_name)
+
+
 class TickTock:
     """
     A taker that works like Matlab's Tic and Toc.
@@ -74,9 +89,11 @@ class TickTock:
     tt.end_profile()
     ```
     """
-    _start = 0
-    profiles: dict[str, Profile] = {}
-    _profile_stack: list[Profile] = list()
+
+    def __init__(self):
+        self._start = 0
+        self.profiles: dict[str, Profile] = {}
+        self._profile_stack: list[Profile] = list()
 
     def tick(self) -> float:
         self._start = perf_counter()
@@ -86,7 +103,7 @@ class TickTock:
         end = perf_counter()
         return end - self._start
 
-    def profile(self, name: str):
+    def profile(self, name: str) -> _ProfileContext:
         """ Begin profile with given name """
         if name not in self.profiles:
             self.profiles[name] = Profile(
@@ -96,25 +113,40 @@ class TickTock:
             )
         self._profile_stack.append(self.profiles[name])
         self.profiles[name].start = perf_counter()
+        return _ProfileContext(self, name)
 
     def end_profile(self, name: str=None) -> float:
         """ End profile. If name given, it is checked that it matches latest started profile
         Return time passed since .profile was called """
         end = perf_counter()
         dt = end - self._profile_stack[-1].start
+        if name is not None and name != self._profile_stack[-1].name:
+            raise NameError(f"Expected to pop profile '{self._profile_stack[-1].name}', received '{name}'")
         self._profile_stack[-1].hits.append(dt)
-        profile = self._profile_stack.pop()
-        if name is not None:
-            assert name == profile.name, f"Expected to pop profile '{profile.name}', received '{name}'"
+        self._profile_stack.pop()
         return dt
 
-    def fuse(self, tt):
+    def fuse(self, tt: TickTock):
         """ Fuses a TickTock instance into self """
+        if len(self._profile_stack) or len(tt._profile_stack):
+            raise ValueError("Unable to fuse while some profiles are still unfinished")
         for profile in tt.profiles.values():
-            if profile.name in self.profiles.keys():
-                self.profiles[profile.name]._hits += profile.hits
+            existing = self.profiles.get(profile.name)
+            if existing:
+                existing._hits += profile.hits
             else:
                 self.profiles[profile.name] = profile
+
+    @staticmethod
+    def fuse_multiple(tts: list[TickTock]) -> TickTock:
+        """ Combine multiple TickTocks """
+        ticktock = TickTock()
+        ids = set(id(tt) for tt in tts)
+        if len(ids) < len(tts):
+            raise ValueError("Some TickTocks are the same instance, which is not allowed")
+        for tt in tts:
+            ticktock.fuse(tt)
+        return ticktock
 
     def remove_outliers(self, threshold=2):
         """ For all profiles, remove hits longer than threshold * average hit """