+++ HISTORY.md
@@ -3,6 +3,9 @@
 ## 0.5.5
 
 - Test logging now uses `Levels.DEBUG` by default
+- Added `TickTock.fuse_multiple` for combining several `TickTock` instances
+- Fixed bugs when using multiple `TickTock` instances
+- Now possible to profile using `with` statement
 
 ## 0.5.4 - Breaking changes